requirements.txt
absl-py==2.1.0
accelerate>=1.0.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
bitsandbytes==0.42.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
datasets>=3.0.1
debugpy==1.8.6
decorator==5.1.1
diffusers==0.30.3
dill==0.3.8
einops==0.8.0
executing==2.1.0
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
gradio==5.1.0
gradio_client==1.4.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub>=0.25.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
libclang==18.1.1
loralib==0.1.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib>=3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mlx>=0.18.1
mlx-lm>=0.19.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy>=1.26.4
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft @ git+https://github.com/huggingface/peft.git@a724834ac43b9478b066d3ec8b421489151f3815
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML>=6.0.2
pyyaml>=6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests>=2.32.3
rich>=13.9.2
ruff==0.6.9
rouge-score>=0.1.2
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
sentencepiece>=0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tokenizers>=0.20.0
tomlkit==0.12.0
torch==2.4.1
tornado==6.4.1
tqdm>=4.66.5
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@698b36da72ae8377fb08ade92b131069898007c2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.14.0
zipp==3.20.2
